refactor scenes directory structure
