+++ src/scene-manager.ts
@@ -1,5 +1,5 @@
 import core from './enchant/core';
-import { Scenes, Scene, SceneKind } from './scenes';
+import { Scenes, Scene, SceneKind } from './scenes/scenes';
 import TopScene from './scenes/top-scene';
 import PlayingScene from './scenes/playing-scene';
 import { code } from './blockly-main';
